Revert "Remove negative bearings on 2048-em glyphs"
This reverts commit 59c45ba4eff539d2a83a7d7004f056b9860d3253.

[why]
The commit breaks the non-mono Nerd Fonts for a lot of people.
The issue was not very good documented and investigation can not be
seen.

For the TITLE it should have affected only the 2048-em Symbols only
font, but in fact all patched fonts were changed. Maybe that was
intended, maybe not.

[how]
This will make the advance width again equal for all glyphs.
This enables the use of the non-mono variant in more terminals.
For wider symbols a space is now (again) needed after the symbol.
That is expected by a lot applications.

[note]
See Pull Request 764 for more details.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>