font-patcher: Increase overlap for critical glyphs
[why]
With the previous commit the gap between powerline glyphs became much
better, but a faint line is still visible for LCD antialiasing (on my
machine).

Having seen how big the overlap is for Cascadia Code the overlap here
can be increase maybe a bit.

[how]
Increase the overlap to 3 % (from mostly 2 %) for the glyphs that have a
'full colored edge' on one side.

Fixes: #29

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>